+++ CHANGELOG.md
@@ -1,5 +1,6 @@
 # Flextype 0.7.3, 2018-12-12
-* Use Imagine library for image manipulation
+* Settings merged into one settings.yaml file
+* Using Imagine library for image manipulation
 * Flextype Component - I18n updated to 1.2.0
 * Flextype Component - Filesystem updated to 1.1.3
 * Symfony YAML - updated to 4.2.1